Move user.sh to last to be loaded, so it can override default definitions
If a user wants to override a setting (let's say, PATH environment variable), it is not currently possible because it is overwritten by some other shell script along the load process. Making user.sh the last to be loaded, the user get's this freedom without having future updates to break their own configuration.